+++ resources/synth/test.js
@@ -2,86 +2,41 @@ var test = {
 	"module": {
 		"name": "net",
 		"path": "d/project/FPGA/FPGA_Design/TCL_project/Test/sim_test/user/src/Verilog/template.v",
-		"text": `
-        // template
-module  template  #(
-  parameter    INPUT_WIDTH  = 12,
-  parameter    OUTPUT_WIDTH = 12
-)(
-    input [INPUT_WIDTH 
-    - 1 : 0]data_in[4:3],
-    output reg clk_in = (INPUT_WIDTH - 
-    OUTPUT_WIDTH) ,
-      clk=9'hd0,
-    input rst_n, RST,
-    output [OUTPUT_WIDTH - 1 : 0] data_out
-);
-  
-endmodule  //template
+		"text": `module up3down5(clock, data_in, up, down, carry_out, borrow_out, count_out, parity_out);
 
-
-module test # (
-    parameter    INPUT_WIDTH  = 12,
-    parameter    OUTPUT_WIDTH = 12
-)(
-    input                           clk_in,
-    input                           rst_n,
-    input  [INPUT_WIDTH - 1 : 0]    data_in , [3:2] dasta_ff,
-    output reg signed [OUTPUT_WIDTH - 1 : 0]   data_out,
-    reg signed [OUTPUT_WIDTH - 1 : 0]   data_ff[5:2]
-);
-  
-wire               	valid_out;
-
-Cordic #(
-	.XY_BITS      		( 12       		),
-	.PH_BITS      		( 32       		),
-	.ITERATIONS   		( 32       		),
-	.CORDIC_STYLE 		( "ROTATE" 		),
-	.PHASE_ACC    		( "ON"     		))
-u_Cordic(
-	//input
-	.clk_in    		( clk_in    		),
-	.RST       		( RST       		),
-	.x_i       		( x_i       		),
-	.y_i       		( y_i       		),
-	.phase_in  		( phase_in  		),
-	.valid_in  		( valid_in  		),
-
-	//output
-	.x_o       		( x_o       		),
-	.y_o       		( y_o       		),
-	.phase_out 		( phase_out 		),
-	.valid_out 		( valid_out 		)
-
-	//inout
-);
-
-wire [3 : 0] 	count_high;
-wire [3 : 0] 	count_low;
-wire              	over;
-
-template u_template(
-	//input
-	.clk        		( clk        		),
-	.data       		( data       		),
-	.en         		( en         		),
-	.load       		( load       		),
-	.rst        		( rst        		),
-	.switch     		( switch     		),
-
-	//output
-	.count_high 		( count_high 		),
-	.count_low  		( count_low  		),
-	.over       		( over       		)
-
-	//inout
-);
-
-
-
-endmodule  //test
-        `
+        input [8:0] data_in;
+        input clock, up, down;
+        
+        output reg [8:0] count_out;
+        output reg carry_out, borrow_out, parity_out;
+        
+        reg [9:0] cnt_up, cnt_dn;
+        reg [8:0] count_nxt;
+        
+        always @(posedge clock) begin
+            cnt_dn = count_out - 3'b 101;
+            cnt_up = count_out + 2'b 11;
+        
+            case ({up,down})
+                2'b 00 :
+                    count_nxt = data_in;
+                2'b 01 :
+                    count_nxt = cnt_dn;
+                2'b 10 :
+                    count_nxt = cnt_up;
+                2'b 11 :
+                    count_nxt = count_out;
+                default :
+                    count_nxt = 9'bX;
+            endcase
+        
+            parity_out  <= ^count_nxt;
+            carry_out   <= up & cnt_up[9];
+            borrow_out  <= down & cnt_dn[9];
+            count_out   <= count_nxt;
+        end
+        
+        endmodule`
 	},
 	"instInfo": [],
 	"includeInfo": []